sepik_fresh/requirements.txt
flask>=3.0.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-flask>=1.2.0
gunicorn>=21.2.0
